% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrans.R
\name{addTrans}
\alias{addTrans}
\title{addTrans}
\usage{
addTrans(color, trans)
}
\arguments{
\item{color}{Colour for plotting}

\item{trans}{Transparency ranging 0 - 255, 0 being fully transparent and 255
fully visible}
}
\value{
res colour code for direct use in plotting
}
\description{
addTrans function for making plot colours transparent
}
