% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_evol.R
\name{oxygen_evol}
\alias{oxygen_evol}
\title{oxygen_evol.}
\usage{
oxygen_evol(fluorwin_filename, firesting_filename, no_light_steps,
  time_step = 60, calibration_file = NA, data_output = F)
}
\arguments{
\item{fluorwin_filename}{name of the fluorwin .txt file to process.}

\item{firesting_filename}{name of the firesting oxygen data file to process.}

\item{no_light_steps}{the number of light steps in the curve.}

\item{time_step}{length of the window (in seconds) of oxygen data to extract per
light step.}

\item{calibration_file}{flashlet calibration file used to calculate PAR from
voltages for FL3500 data. Generic calibration file is built into package
so leave as NA if no specific calibration available.}

\item{data_output}{logical (default FALSE), if TRUE dataframes produced are
written to CSV and plots written to single .pdf file (see Values below)}
}
\value{
oxy_results list containing the following two dataframes:
  1. raw_oxygen dataframe containing the tidy outputs from the Firesting
  optode (i.e. complete oxygen trace with formatted time columns).
  2. matched_oxygen dataframe containing the FL3500 light step with matched
  calculated rate of oxygen evolution (units of umol s-1), with columns
  showing voltage and PAR levels per light step

  If data_output = TRUE:
  x2 CSV files of raw_oxygen and matched_oxygen datasets appended with oxygen
  datafile name

  AND

  .pdf document appended with oxygen datafile name showing:
  1. complete oxygen trace with grey areas highlighting which sections have
  been clipped and matched to fluorescence data (As determined by the
  time_step selected)
  2. individual plots of each linear regression for oxygen evolution
  calculations
  3. calculated oxygen evolution ~ light step over the course of the light
  curve with PAR levels shown in blue trace in brackground
  4. scatter plot of oxygen evolution ~ PAR
}
\description{
extract and match firesting oxygen data to FL3500 PSI fluorescence data.
}
\examples{
To import example data, run the following code:

example_fluorwin_data<-system.file("extdata", "example_fluorwin_data.txt", package = "OxygenEvol")
example_oxygen_data<-system.file("extdata", "example_oxygen_data.txt", package = "OxygenEvol")
oxygen_evol(example_fluorwin_data, example_oxygen_data, no_light_steps=11, time_step=60, calibration_file=NA)
}
