% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_evol.R
\name{oxygen_evol}
\alias{oxygen_evol}
\title{oxygen_evol.}
\usage{
oxygen_evol(fluorwin_filename, firesting_filename, no_light_steps,
  time_step = 60, calibration_file = NA)
}
\arguments{
\item{fluorwin_filename}{name of the fluorwin .txt file to process.}

\item{firesting_filename}{name of the firesting oxygen data file to process.}

\item{no_light_steps}{the number of light steps in the curve.}

\item{time_step}{length of the window (in seconds) of oxygen data to extract per
light step.}

\item{calibration_file}{flashlet calibration file used to calculate PAR from
voltages for FL3500 data. Generic calibration file is built into package
so leave as NA if no specific calibration available.}
}
\value{
raw_oxygen dataframe containing the tidy outputs from the Firesting
  optode (i.e. complete oxygen trace with formatted time columns).

  matched_oxygen dataframe containing the FL3500 light step with matched
  calculated rate of oxygen evolution (units of umol s-1), with columns
  showing voltage and PAR levels per light step

  a single pdf document (taking its name from the oxygen data input filename)
  showing i) complete oxygen trace with grey areas
  highlighting which sections have been clipped and matched to fluorescence
  data (As determined by the time_step selected), ii) individual plots of each
  linear regression for oxygen evolution calculation, iii) calculated oxygen
  evolution ~ light step over the course of the light curve, iv) scatter plot
  of oxygen evolution ~ PAR (if a sequential light curve is used this will
  show the oxygen evolution over the course of the light curve, if a non
  -sequential
  light curve was used this will show the overall relationship).
}
\description{
extract and match firesting oxygen data to FL3500 PSI fluorescence data.
}
\examples{
oxygen_evol(example_fluorwin_data, example_oxygen_data, no_light_steps=11, time_step=60, calibration_file=NA)
}
